app/requirements.txt
sqlmodel==0.0.27
pydantic_settings